+++ CHANGELOG.md
@@ -8,10 +8,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.8.0] - 2024-07-23
+
 ### Added
 
 - Added bottom sheet Composable APIs(`BottomSheetListContent`, `BottomSheetSliderContent`, etc...).
 - Added bottom sheet item Composable APIs(`BottomSheetCheckboxItem`, `BottomSheetSingleSelectionItem`).
+
+
+## [0.7.5] - 2024-07-10
+
+### Added
+
+- InputOTPField composable is introduced with [TrendyolInputOTPFieldStyle.Outline] and
+  [TrendyolInputOTPFieldStyle.Filled] styles.
+- InputPasswordField composable is introduced with [TrendyolOutlinedTextFieldStyle.Outline],
+  [TrendyolOutlinedTextFieldStyle.FloatingLabelOutlined] and [FloatingLabelOutlined.Filled] styles.
+- OutlinedTextField disabled label color changed to colorOnSurfaceVariant1.
+
+## [0.7.4] - 2024-07-19
+
+### Added
+
+- Add new vertical price api
 
 ## [0.7.3] - 2024-06-20